requirements.txt
playwright==1.40.0
groq==0.4.1
python-dotenv==1.0.0

